requirements.txt
setuptools~=54.2.0
argparse~=1.4.0
opencv-python~=4.5.1.48
tqdm~=4.60.0
numpy~=1.20.2
pytesseract~=0.3.7
Pillow~=8.2.0
matplotlib~=3.4.1
seaborn~=0.11.1
pandas~=1.2.3